cxx/ctrl/mc: append sequence to full run id if split failed

Closes #355

See merge request constellation/constellation!1163

+++ cxx/controllers/MissionControl/MissionControl.hpp
@@ -10,6 +10,7 @@
 #pragma once
 
 #include <cstddef>
+#include <cstdint>
 #include <map>
 #include <optional>
 #include <string>
@@ -176,7 +177,7 @@ private:
      * @param run_id Run identifier string
      * @return Run name and sequence
      */
-    std::pair<std::string, std::size_t> split_run_identifier(const std::string& run_id) const;
+    std::pair<std::string, std::int64_t> split_run_identifier(const std::string& run_id) const;
 
     /**
      * @brief Helper to parse the configuration file for all satellites
